Remove fixture files with Windows incompatible names
Quick fix: Some fixture files for sysfs have names which are invalid on
Windows.  This breaks git checkout on Windows systems. As procfs is a
depdency of client_golang, this breaks Windows builds on hundreds of
projects.

A nicer solution could rename all paths with characters invalid on some
filesystems. Such a rename approach should ideally work with default
tools available on OSX and Linux.