Replace fixtures.tar.gz with regular files
In order to keep the fixtures in a form that can easily be reviewed,
this changeset puts the files from fixtures.tar.gz back into the repo as
regular files (in a separate directory, fixtures.src). In order to
prevent filenames that prevent checkouts on Windows, colons are encoded
as "_@colon@_" and copied to the correct location by make.

Alternatives: files could be renamed rather than copied (using -depth
in the find command to rename files before their parent directories).
However, doing so would remove files from a location where git expects
them.