Feature: Add check_license.sh
As described in issue #37 there is no automated script to check that all
source files contain a license header.

Prometheus already has a script provided by @jonboulle and I have simply
copied that script and applied it to this code base.

Example output:
```
github.com/prometheus/procfs on  feat-add-license-check [+]
➜ make check_license
>> checking license header
-e license header checking failed:
-e   ./fs.go
-e   ./fs_test.go
-e   ./ipvs.go
...
```

All credit to @jonboulle for the initial script.

This closes #37