move bad buddyinfo test data into sources
The fixtures.tar proc and sys directories should only include valid
paths similar to what would be seen on an actual linux system.
This moves the buddyinfo test data into the proper filesystem location
or inline in the test sources in the case of the parser tests.

See issue #43 for more info.

Signed-off-by: Paul Gier <pgier@redhat.com>