Fix parsing of /sys/class/power_supply
The file `/sys/class/power_supply/<power_supply>/device` is a symlink to a
directory, but it is read as a regular file. This commit fixes this problem and
tries to solve #205.

Also it updates the fixtures.ttar file and add the symlink in it to better
represent the situation on the pc.

Signed-off-by: Sven Haardiek <sven@haardiek.de>