introduce new function ReadFileNoStat
This function should be a slight improvement over
ioutil.ReadFile because it skips the call to os.Stat which commonly
returns the wrong file size for files in /proc and /sys.

Usage of this function also makes the code a bit more concise for
parsing of several of the files in /proc.

Signed-off-by: Paul Gier <pgier@redhat.com>