internal/util: Read(U)intFromFile: return file name in errors
This improves error messages by adding the affected file name.
This helps downstream users to locate the source of errors which
is useful for e.g. prometheus/node_exporter#1710.

The original error is wrapped as some consumers perform checks on it.
The consumers within procfs have been updated to Unwrap the error if
necessary.

Signed-off-by: Christian Hoffmann <mail@hoffmann-christian.info>