Parse /sys/class/drm/cardN/device/ stats
This commits adds support for reading card stats exposed by drm.
Only `amdgpu` driver currently exposes these stats but it should be
easy to add more driver support if required.
FreeBSD uses the same driver through `kms-drm` but it does not expose
the attributes we are interested in through niether `sysctl` or `linsysfs`.

Signed-off-by: Siavash Safi <siavash.safi@gmail.com>