Adds test for OmniPath device detection
This adds fixtures and a test for the OmniPath detection and metric
reading added in a previous commit.

This could use a check from someone better acquainted with the driver
and relevant metrics, but it should do some approximation of the job for now.

It might even be sensible to specifically detect major versions of devices, since
the detailed metric file layout might even vary between mlx4 and mlx5.

There's also a factor of 4 included in the Mellanox metric pickup that
I'm not sure about: might need to detect device to tell whether we need
that, and that's a bit more work.

Signed-off-by: Ian Kirker <ian.kirker@gmail.com>