Check whether FS is a real procfs where it matters
For some behaviors there is no substitution for real `/proc`.

One example is `stat()` for `/proc/<pid>/fd` returning the number
of open files for `procfs`. The field has a different meaning on
regular filesystems and we need to check for it in tests.

Work around this by carrying the fact that `FS` is real in the `FS`
struct itself, so that it can be checked where it matters. The realness
is determined once per `FS` creation by doing `statfs()` and checking
for the `PROC_SUPER_MAGIC` match in the returned fs type.

Signed-off-by: Ivan Babrou <github@ivan.computer>