Custom build tag "nostatfs" disables statfs
This allows users to specify the "nostatfs" build tag to disable this feature if needed.

In our case, we are using Tamago (https://github.com/usbarmory/tamago), which does not support syscall.Statfs. Without this change attempting to build returns 'undefined: syscall.Statfs' and 'undefined: syscall.Statfs_t'. With this change, we can build successfully build our tamago executables provided that "nostatfs" is added to the build tags.

This option seems preferable to adding "tamago" as a build tag here, and will scale to other use cases too.

Signed-off-by: Martin Hutchinson <mhutchinson@gmail.com>