Fix Proc.Limits limit name matching
I was working on improving this algorithm to reduce the number of
allocations when I found out that with the addition of the additional
test cases, `Max processes` was failing to match the `switch`
statement as for some reason the limit name has a trailing
whitespace. By trimming the spaces it now matches all cases.

Signed-off-by: Leandro López (inkel) <leandro.lopez@grafana.com>